Original change: https://googleplex-android-review.googlesource.com/c/platform/frameworks/base/+/14824025

+++ core/java/android/os/incremental/IncrementalMetrics.java
@@ -90,7 +90,7 @@ public class IncrementalMetrics {
      * @return total duration in milliseconds of delayed reads
      */
     public long getTotalDelayedReadsDurationMillis() {
-        return mData.getInt(IIncrementalService.METRICS_TOTAL_DELAYED_READS_MILLIS, -1);
+        return mData.getLong(IIncrementalService.METRICS_TOTAL_DELAYED_READS_MILLIS, -1);
     }
 
     /**
